Create card.js, modal.js
